==> 04d_selecao_variaveis_regionais.Rmd <==
---
title: "R Notebook"
output: html_notebook
---



```{r library, echo=FALSE, warning=FALSE, message=FALSE}
library(tidyverse)
library(xlsx)
```

```{r df publico}
df_publico <- read.csv2("output/books/df_publico.csv")
df_publico <- df_publico %>% 
  filter(id_dependencia_adm == 'Estadual') %>% 
  select(id_serie,
         id_escola,
         id_municipio,
         media)
head(df_publico)
```


# Seade
```{r seade 2010}
seade2010 <- read_csv2("data/regionais/seade_2010.csv", locale(encoding = "ISO-8859-1"), col_names = TRUE, col_types = NULL)
colnames(seade2010) <- c("no_municipio","ano", "idh_longevidade_2010_imp", "coleta_lixo_2010_imp", "esgoto_2010_imp", "co_municipio")
head(seade2010)
```

```{r seade 2019}
seade2019 <- read_csv2("data/regionais/seade_2019.csv", locale(encoding = "ISO-8859-1"), col_names = TRUE, col_types = "cddddd")
colnames(seade2019) <- c("no_municipio","ano", "nascidos_vivos_mae18_2019_imp", "rendimento_medio_2019_imp", "ind_envelhecimento_2019_imp", "co_municipio")
head(seade2019)


```

```{r colnames 2010}
names_2010 <- colnames(seade2010)
names_2010
```

```{r colnames 2019}
names_2019 <- colnames(seade2019)
names_2019
```

## Gera base por série

```{r 5Ef}
seade5ef <- df_publico %>% 
  filter(id_serie == '5EF') %>% 
  left_join(seade2010[,names_2010[4:6]], by = c("id_municipio" = "co_municipio")) %>%  
  left_join(seade2019[,names_2019[4:6]], by = c("id_municipio" = "co_municipio"))

head(seade5ef) 
```

```{r 9Ef}
seade9ef <- df_publico %>% 
  filter(id_serie == '9EF') %>% 
  left_join(seade2010[,names_2010[3:6]], by = c("id_municipio" = "co_municipio")) %>%  
  left_join(seade2019[,names_2019[5:6]], by = c("id_municipio" = "co_municipio"))

head(seade9ef) 
```

```{r 3EM}
seade3em <- df_publico %>% 
  filter(id_serie == '3EM') %>% 
  left_join(seade2010[,names_2010[3:6]], by = c("id_municipio" = "co_municipio")) %>%  
  left_join(seade2019[,names_2019[c(3,5,6)]], by = c("id_municipio" = "co_municipio"))

head(seade3em) 
```
```{r}
seade5ef <- seade5ef %>% select(-id_escola,-media) %>% unique()
seade9ef <- seade9ef %>% select(-id_escola,-media) %>% unique()
seade3em <- seade3em %>% select(-id_escola,-media) %>% unique()
```



```{r Export}
write.csv2(seade5ef, "output/books/seade5ef.csv", row.names = FALSE, na = "")
write.csv2(seade9ef, "output/books/seade9ef.csv", row.names = FALSE, na = "")
write.csv2(seade3em, "output/books/seade3em.csv", row.names = FALSE, na = "")
```
