==> 08_cria_pilares_5ef.Rmd <==
---
title: "R Notebook"
output: html_notebook
params:
  serie: '5ef'
---


```{r}
library(tidyverse)
library(caret)
library(rpart)
library(rpart.plot)

```

# Import 
```{r modelo v1}
url <- paste0("output/v1/iee_", params$serie, "_v1.csv")
df <- read.csv2(url, na = "") 
df <- df %>% 
  select(id_escola,
         co_distrito,
         id_municipio,
         media)
head(df)
```

```{r importa saeb full}
url <- paste0("output/books/saeb", params$serie, "_prop.csv")
saeb_full <- read.csv2(url)

saeb_select <- saeb_full %>% 
  select(id_escola,
    starts_with("tx_resp_q003"), #família
    starts_with("tx_resp_q004"), #letramento
    starts_with("tx_resp_q005"), #letramento
    starts_with("tx_resp_q008"), #região
    starts_with("tx_resp_q011"), #localização
    starts_with("tx_resp_q012"), #localização
    starts_with("tx_resp_q009"), #bens
    starts_with("tx_resp_q010"), #bens
    starts_with("tx_resp_q007")  #bens
    )
```

```{r importa ipvs}
url <- paste0("output/books/ipvs", params$serie, ".csv")
ipvs <- read.csv2(url, na = "") 
```

```{r importa seade}
url <- paste0("output/books/seade", params$serie, ".csv")
seade <- read.csv2(url, na = "")
```

```{r importa saresp}
url <- paste0("output/books/saresp", params$serie, "_prop.csv")
saresp <- read.csv2(url, na = "")
```

# Integra
```{r}
df <- df %>% 
  left_join(saeb_select, by = "id_escola") %>% 
  left_join(saresp, by = "id_escola") %>% 
  left_join(ipvs, by = "co_distrito") %>% 
  left_join(seade, by = "id_municipio")

head(df)
```

# Pilares
## Pilar Família
```{r}
rpart_familia <- rpart(media ~ tx_resp_q003a_B + tx_resp_q003b_B + tx_resp_q003c_B + tx_resp_q003d_B + tx_resp_q003e_B,
                       data = df)
rpart.plot(rpart_familia)
```


### Cria Pilar
```{r}
df <-  df %>% 
  mutate(pred_familia = round(predict(rpart_familia, df)),
         pilar_familia = case_when(
           pred_familia == 249 ~ 'A',
           pred_familia == 240 ~ 'B',
           pred_familia == 237 ~ 'C',
           pred_familia %in% c(229,228) ~ 'D',
           pred_familia == 217 ~ 'E',
         ))

table(df$pilar_familia)
```
### R2
```{r}
lm_familia <- lm(media ~ pilar_familia, df)
summary(lm_familia)$r.squared
```

## Pilar Letramento
### Rpart
```{r}
rpart_letramento <- rpart(media ~ .,
                       data = df %>% select(media,starts_with("tx_resp_q004"),starts_with("tx_resp_q005"))
                       ,minbucket = 0.05*nrow(df)
                       )
rpart.plot(rpart_letramento)
```

### Cria Pilar
```{r}
df <-  df %>% 
  mutate(pred_letramento = round(predict(rpart_letramento, df)),
         pilar_letramento = case_when(
           pred_letramento == 249 ~ 'A',
           pred_letramento == 240 ~ 'B',
           pred_letramento == 231 ~ 'C',
           pred_letramento == 224 ~ 'D',
         ))

table(df$pilar_letramento)
```

### R2
```{r}
model <- lm(media ~ pilar_letramento, df)
summary(model)$r.squared
```



## Pilar Região
### Rpart
```{r}
rpart_regiao <- rpart(media ~ tx_resp_q008a_B + tx_resp_q008b_B + tx_resp_q008c_B,
                       data = df
                       ,minbucket = 0.05*nrow(df)
                       )
rpart.plot(rpart_regiao)
```

### Cria Pilar
```{r}
df <-  df %>% 
  mutate(pred_regiao = round(predict(rpart_regiao, df)),
         pilar_regiao = case_when(
           pred_regiao == 247 ~ 'A',
           pred_regiao == 240 ~ 'B',
           pred_regiao == 234 ~ 'C',
           pred_regiao == 229 ~ 'D',
           pred_regiao == 224 ~ 'E',
         ))

table(df$pilar_regiao)
```

### R2
```{r}
model <- lm(media ~ pilar_regiao, df)
summary(model)$r.squared
```




## Pilar Localização
### Rpart
```{r}
rpart_localizacao <- rpart(media ~ .,
                       data = df %>% select(media,starts_with('tx_resp_q011'),starts_with('tx_resp_q012'))
                       ,minbucket = 0.05*nrow(df)
                       )
rpart.plot(rpart_localizacao)
```

### Cria Pilar
```{r}
df <-  df %>% 
  mutate(pred_localizacao = round(predict(rpart_localizacao, df)),
         pilar_localizacao = case_when(
           pred_localizacao == 254 ~ 'A',
           pred_localizacao == 245 ~ 'B',
           pred_localizacao == 237 ~ 'C',
           pred_localizacao == 235 ~ 'C',
           pred_localizacao == 230 ~ 'D',
           pred_localizacao == 227 ~ 'E',
           pred_localizacao == 219 ~ 'E'
         ))

table(df$pilar_localizacao)
```

### R2
```{r}
model <- lm(media ~ pilar_localizacao, df)
summary(model)$r.squared
```




## Pilar Bens 
### Rpart
```{r}
rpart_bens <- rpart(media ~ .,
                       data = df %>% select(media,starts_with('tx_resp_q009'),starts_with('tx_resp_q010'))
                       ,minbucket = 0.05*nrow(df)
                       )
rpart.plot(rpart_bens)
```

### Cria Pilar
```{r}
df <-  df %>% 
  mutate(pred_bens = round(predict(rpart_bens, df)),
         pilar_bens = case_when(
           pred_bens == 253 ~ 'A',
           pred_bens == 243 ~ 'B',
           pred_bens == 242 ~ 'B',
           pred_bens == 235 ~ 'C',
           pred_bens == 233 ~ 'C',
           pred_bens == 227 ~ 'D',
           pred_bens == 225 ~ 'D',
           pred_bens == 216 ~ 'E',
         ))

table(df$pilar_bens)
```

### R2
```{r}
model <- lm(media ~ pilar_bens, df)
summary(model)$r.squared
```




## Pilar Bens Supérfluos
### Rpart
```{r}
rpart_bens <- rpart(media ~ .,
                       data = df %>% select(media,
                                            starts_with('tx_resp_q009b'),
                                            starts_with('tx_resp_q009c'),
                                            starts_with('tx_resp_q009e'),
                                            starts_with('tx_resp_q009g'))
                       ,minbucket = 0.05*nrow(df)
                       )
rpart.plot(rpart_bens)
```

### Cria Pilar
```{r}
df <-  df %>% 
  mutate(pred_bens = round(predict(rpart_bens, df)),
         pilar_bens_superfluos = case_when(
           pred_bens >= 254 ~ 'A',
           pred_bens >= 242 ~ 'B',
           pred_bens >= 231 ~ 'C',
           pred_bens >= 224 ~ 'D',
           pred_bens >= 217 ~ 'E'
         ))

table(df$pilar_bens_superfluos)
```

### R2
```{r}
model <- lm(media ~ pilar_bens_superfluos, df)
summary(model)$r.squared
```




## Pilar Bens serviços
### Rpart
```{r}
rpart_bens <- rpart(media ~ .,
                       data = df %>% select(media,
                                            starts_with('tx_resp_q007'),
                                            starts_with('tx_resp_q010a'),
                                            starts_with('tx_resp_q010b'))
                       ,minbucket = 0.05*nrow(df)
                       )
rpart.plot(rpart_bens)
```

### Cria Pilar
```{r}
df <-  df %>% 
  mutate(pred_bens = round(predict(rpart_bens, df)),
         pilar_bens_servicos = case_when(
           pred_bens >= 250 ~ 'A',
           pred_bens >= 240 ~ 'B',
           pred_bens >= 233 ~ 'C',
           pred_bens >= 227 ~ 'D',
           pred_bens >= 216 ~ 'E',
         ))

table(df$pilar_bens_servicos)
```

### R2
```{r}
model <- lm(media ~ pilar_bens_servicos, df)
summary(model)$r.squared
```





## Pilar Bens Cômodos
### Rpart
```{r}
rpart_bens <- rpart(media ~ .,
                       data = df %>% select(media,
                                            starts_with('tx_resp_q009d'),
                                            starts_with('tx_resp_q009f'),
                                            starts_with('tx_resp_q010c'),
                                            starts_with('tx_resp_q010d'),
                                            starts_with('tx_resp_q010e'))
                       ,minbucket = 0.05*nrow(df)
                       )
rpart.plot(rpart_bens)
```

### Cria Pilar
```{r}
df <-  df %>% 
  mutate(pred_bens = round(predict(rpart_bens, df)),
         pilar_bens_comodos = case_when(
           pred_bens >= 250 ~ 'A',
           pred_bens >= 243 ~ 'B',
           pred_bens >= 235 ~ 'C',
           pred_bens >= 230 ~ 'D',
           pred_bens >= 226 ~ 'E',
           pred_bens >= 219 ~ 'F',
         ))

table(df$pilar_bens_comodos)
```

### R2
```{r}
model <- lm(media ~ pilar_bens_comodos, df)
summary(model)$r.squared
```





## Pilar Bens Eletrodomésticos
### Rpart
```{r}
rpart_bens <- rpart(media ~ .,
                       data = df %>% select(media,
                                            starts_with('tx_resp_q009a'),
                                            starts_with('tx_resp_q010f'),
                                            starts_with('tx_resp_q010g'),
                                            starts_with('tx_resp_q010h'),
                                            starts_with('tx_resp_q010i'))
                       ,minbucket = 0.05*nrow(df)
                       )
rpart.plot(rpart_bens)
```

### Cria Pilar
```{r}
df <-  df %>% 
  mutate(pred_bens = round(predict(rpart_bens, df)),
         pilar_bens_eletrodomesticos = case_when(
           pred_bens >= 253 ~ 'A',
           pred_bens >= 242 ~ 'B',
           pred_bens >= 233 ~ 'C',
           pred_bens >= 224 ~ 'D',
           pred_bens >= 216 ~ 'E',
         ))

table(df$pilar_bens_eletrodomesticos)
```

### R2
```{r}
model <- lm(media ~ pilar_bens_eletrodomesticos, df)
summary(model)$r.squared
```

# Pilar IPVS
### Rpart
```{r}
rpart_ipvs <- rpart(media ~ .,
                       data = df %>% select(media,
                                            ends_with('ipvs'))
                       ,minbucket = 0.05*nrow(df)
                       )
rpart.plot(rpart_ipvs)
```

```{r}
df <-  df %>% 
  mutate(pred_ipvs = round(predict(rpart_ipvs, df)),
         pilar_ipvs = case_when(
           pred_ipvs >= 242 ~ 'A',
           pred_ipvs >= 232 ~ 'B',
           pred_ipvs >= 226 ~ 'C'
         ))

table(df$pilar_ipvs)
```
### R2
```{r}
model <- lm(media ~ pilar_ipvs, df)
summary(model)$r.squared
```

# Pilar Seade
### Rpart
```{r}
rpart_seade <- rpart(media ~ .,
                       data = df %>% select(media,
                                            ends_with('imp'))
                       ,minbucket = 0.05*nrow(df)
                       )
rpart.plot(rpart_seade)
```
```{r}
df <-  df %>% 
  mutate(pred_seade = round(predict(rpart_seade, df)),
         pilar_seade = case_when(
           pred_seade >= 241 ~ 'A',
           pred_seade >= 236 ~ 'B',
           pred_seade >= 231 ~ 'C'
         ))

table(df$pilar_seade)
```
### R2
```{r}
model <- lm(media ~ pilar_seade, df)
summary(model)$r.squared
```

# Pilar Saresp
### Rpart
```{r}
df <- df %>% mutate_at(vars(q21_A_spp:q50_E_spp),~ifelse(is.na(.x), mean(.x, na.rm = TRUE), .x))
#colSums(is.na(df %>% select(ends_with('spp'))))

df <- df %>% 
  mutate(
    q23_EFG_spp = q23_E_spp + q23_F_spp + q23_G_spp)

rpart_saresp <- rpart(media ~ .,
                       data = df %>% select(media,
                                            ends_with('spp'))
                       ,minbucket = 0.05*nrow(df)
                       )
rpart.plot(rpart_saresp)
```

```{r}

df <-  df %>% 
  mutate(pred_saresp = round(predict(rpart_saresp, df)),
         pilar_saresp = case_when(
           pred_saresp >= 256 ~ 'A',
           pred_saresp >= 247 ~ 'B',
           pred_saresp >= 241 ~ 'C',
           pred_saresp >= 234 ~ 'D',
           pred_saresp >= 231 ~ 'E',
           pred_saresp >= 226 ~ 'F',
           pred_saresp >= 218 ~ 'G',
         ))

table(df$pilar_saresp)
```
### R2
```{r}
model <- lm(media ~ pilar_saresp, df)
summary(model)$r.squared
```

# Export
```{r}
pilar <- df %>% 
  select(id_escola,
         starts_with('pilar'))

head(pilar)
```

```{r}
url <- paste0("output/pilares/pilar_",params$serie,".csv")
write.csv2(pilar, url, row.names = F, na = "")
```
